requirements.txt
# ============================================================
# Personal Finance Manager - AI Insights
# Requirements File
# Python 3.11+
# ============================================================

# --- Web Framework ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
jinja2>=3.1.3
aiofiles>=23.2.1

# --- Database & ORM ---
sqlalchemy>=2.0.25
alembic>=1.13.1

# --- Settings & Environment ---
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# --- ML & Data Science ---
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.3
joblib>=1.3.2

# --- NLP / Embeddings ---
sentence-transformers>=2.4.0

# --- OCR & PDF ---
pytesseract>=0.3.10
pdfplumber>=0.10.3
pillow>=10.2.0

# --- Visualization ---
matplotlib>=3.8.2
plotly>=5.18.0

# --- HTTP Client ---
httpx>=0.26.0

# --- Utilities ---
python-dateutil>=2.8.2
faker>=22.0.0
